+++ tests/unit/_experimental/dags/predictions/test_datarobot_s3_batch_scoring_dag.py
@@ -1,4 +1,4 @@
-# Copyright 2022 DataRobot, Inc. and its affiliates.
+# Copyright 2023 DataRobot, Inc. and its affiliates.
 #
 # All rights reserved.
 #
@@ -7,20 +7,23 @@
 # Released under the terms of DataRobot Tool and Utility Agreement.
 import pytest
 
-from datarobot_provider.example_dags.deprecated.datarobot_score_dag import datarobot_score
+from datarobot_provider._experimental.example_dags.datarobot_aws_s3_batch_scoring_dag import (
+    datarobot_s3_batch_scoring,
+)
 
 
 def test_dag_loaded(dagbag):
-    dag = dagbag.get_dag(dag_id="datarobot_score")
+    dag = dagbag.get_dag(dag_id="datarobot_s3_batch_scoring")
     assert dagbag.import_errors == {}
     assert dag is not None
-    assert len(dag.tasks) == 2
+    assert len(dag.tasks) == 3
 
 
 def test_dag_structure():
-    dag = datarobot_score()
+    dag = datarobot_s3_batch_scoring()
     pytest.helpers.assert_dag_dict_equal(
         {
+            "get_aws_credentials": ["score_predictions"],
             "score_predictions": ["check_scoring_complete"],
             "check_scoring_complete": [],
         },
